+++ src/pat/inject/inject.js
@@ -1,12 +1,12 @@
+import "../../core/jquery-ext"; // for :scrollable for autoLoading-visible
 import "regenerator-runtime/runtime"; // needed for ``await`` support
 import $ from "jquery";
 import _ from "underscore";
 import ajax from "../ajax/ajax";
-import Parser from "../../core/parser";
 import logging from "../../core/logging";
+import Parser from "../../core/parser";
 import registry from "../../core/registry";
 import utils from "../../core/utils";
-import "../../core/jquery-ext"; // for :scrollable for autoLoading-visible
 
 const log = logging.getLogger("pat.inject");
 const parser = new Parser("inject");
@@ -52,12 +52,10 @@ const inject = {
     name: "inject",
     trigger:
         ".raptor-ui .ui-button.pat-inject, a.pat-inject, form.pat-inject, .pat-subform.pat-inject",
-    init: function inject_init($el, opts) {
-        var cfgs = inject.extractConfig($el, opts);
+    init($el, opts) {
+        const cfgs = this.extractConfig($el, opts);
         if (
-            cfgs.some(function (e) {
-                return e.history === "record";
-            }) &&
+            cfgs.some((e) => e.history === "record") &&
             !("pushState" in history)
         ) {
             // if the injection shall add a history entry and HTML5 pushState
@@ -88,72 +86,27 @@ const inject = {
             }
         }
         if (cfgs[0].pushMarker) {
-            $("body").on("push", function (event, data) {
-                console.log("received push message: " + data);
+            $("body").on("push", (event, data) => {
+                log.debug("received push message: " + data);
                 if (data == cfgs[0].pushMarker) {
-                    console.log("re-injecting " + data);
-                    inject.onTrigger.apply($el[0], []);
+                    log.debug("re-injecting " + data);
+                    this.onTrigger({ target: $el[0] });
                 }
             });
         }
         if (cfgs[0].idleTrigger) {
-            // XXX TODO: handle item removed from DOM
-            var timeout = parseInt(cfgs[0].idleTrigger, 10);
-            var timer;
-
-            const onTimeout = () => {
-                inject.onTrigger.apply($el[0], []);
-                unsub();
-                clearTimeout(timer);
-            };
-
-            var onInteraction = utils.debounce(function onInteraction() {
-                clearTimeout(timer);
-                timer = setTimeout(onTimeout, cfgs[0].trigger);
-            }, timeout);
-
-            const unsub = () => {
-                ["scroll", "resize"].forEach(function (e) {
-                    window.removeEventListener(e, onInteraction);
-                });
-                [
-                    "click",
-                    "keypress",
-                    "keyup",
-                    "mousemove",
-                    "touchstart",
-                    "touchend",
-                ].forEach(function (e) {
-                    document.removeEventListener(e, onInteraction);
-                });
-            };
-
-            onInteraction();
-
-            ["scroll", "resize"].forEach(function (e) {
-                window.addEventListener(e, onInteraction);
-            });
-            [
-                "click",
-                "keypress",
-                "keyup",
-                "mousemove",
-                "touchstart",
-                "touchend",
-            ].forEach(function (e) {
-                document.addEventListener(e, onInteraction);
-            });
+            this._initIdleTrigger($el, cfgs[0].idleTrigger);
         } else {
             switch (cfgs[0].trigger) {
                 case "default":
-                    cfgs.forEach(function (cfg) {
+                    cfgs.forEach((cfg) => {
                         if (cfg.delay) {
                             cfg.processDelay = cfg.delay;
                         }
                     });
                     // setup event handlers
                     if ($el.is("form")) {
-                        $el.on("submit.pat-inject", inject.onTrigger)
+                        $el.on("submit.pat-inject", this.onTrigger.bind(this))
                             .on(
                                 "click.pat-inject",
                                 "[type=submit]",
@@ -162,32 +115,32 @@ const inject = {
                             .on(
                                 "click.pat-inject",
                                 "[type=submit][formaction], [type=image][formaction]",
-                                inject.onFormActionSubmit
+                                this.onFormActionSubmit.bind(this)
                             );
                     } else if ($el.is(".pat-subform")) {
                         log.debug("Initializing subform with injection");
                     } else {
-                        $el.on("click.pat-inject", inject.onTrigger);
+                        $el.on("click.pat-inject", this.onTrigger.bind(this));
                     }
                     break;
                 case "autoload":
                     if (!cfgs[0].delay) {
-                        inject.onTrigger.apply($el[0], []);
+                        this.onTrigger({ target: $el[0] });
                     } else {
                         // generate UID
-                        var uid = Math.random().toString(36);
+                        const uid = Math.random().toString(36);
                         $el.attr("data-pat-inject-uid", uid);
 
                         // function to trigger the autoload and mark as triggered
-                        const delayed_trigger = (uid) => {
+                        const delayed_trigger = (uid_) => {
                             // Check if the element has been removed from the dom
-                            var still_there = $(
-                                "[data-pat-inject-uid='" + uid + "']"
+                            const still_there = $(
+                                "[data-pat-inject-uid='" + uid_ + "']"
                             );
                             if (still_there.length == 0) return false;
 
                             $el.data("pat-inject-autoloaded", true);
-                            inject.onTrigger.apply($el[0], []);
+                            this.onTrigger({ target: $el[0] });
                             return true;
                         };
                         window.setTimeout(
@@ -197,10 +150,10 @@ const inject = {
                     }
                     break;
                 case "autoload-visible":
-                    inject._initAutoloadVisible($el, cfgs);
+                    this._initAutoloadVisible($el, cfgs);
                     break;
                 case "idle":
-                    inject._initIdleTrigger($el, cfgs[0].delay);
+                    this._initIdleTrigger($el, cfgs[0].delay);
                     break;
             }
         }
@@ -209,55 +162,55 @@ const inject = {
         return $el;
     },
 
-    destroy: function inject_destroy($el) {
+    destroy($el) {
         $el.off(".pat-inject");
         $el.data("pat-inject", null);
         return $el;
     },
 
-    onTrigger: function inject_onTrigger(ev) {
+    onTrigger(e) {
         /* Injection has been triggered, either via form submission or a
          * link has been clicked.
          */
-        var cfgs = $(this).data("pat-inject"),
-            $el = $(this);
+        const $el = $(e.target);
+        const cfgs = $el.data("pat-inject");
         if ($el.is("form")) {
-            $(cfgs).each(function (i, v) {
+            $(cfgs).each((i, v) => {
                 v.params = $.param($el.serializeArray());
             });
         }
-        ev && ev.preventDefault();
+        e.preventDefault && e.preventDefault();
         $el.trigger("patterns-inject-triggered");
-        inject.execute(cfgs, $el);
+        this.execute(cfgs, $el);
     },
 
-    onFormActionSubmit: function inject_onFormActionSubmit(ev) {
-        ajax.onClickSubmit(ev); // make sure the submitting button is sent with the form
+    onFormActionSubmit(e) {
+        ajax.onClickSubmit(e); // make sure the submitting button is sent with the form
 
-        var $button = $(ev.target),
-            formaction = $button.attr("formaction"),
-            $form = $button.parents(".pat-inject").first(),
-            opts = { url: formaction },
-            $cfg_node = $button.closest("[data-pat-inject]"),
-            cfgs = inject.extractConfig($cfg_node, opts);
+        const $button = $(e.target);
+        const formaction = $button.attr("formaction");
+        const $form = $button.parents(".pat-inject").first();
+        const opts = { url: formaction };
+        const $cfg_node = $button.closest("[data-pat-inject]");
+        const cfgs = this.extractConfig($cfg_node, opts);
 
-        $(cfgs).each(function (i, v) {
+        $(cfgs).each((i, v) => {
             v.params = $.param($form.serializeArray());
         });
 
-        ev.preventDefault();
+        e.preventDefault();
         $form.trigger("patterns-inject-triggered");
-        inject.execute(cfgs, $form);
+        this.execute(cfgs, $form);
     },
 
-    submitSubform: function inject_submitSubform($sub) {
+    submitSubform($sub) {
         /* This method is called from pat-subform
          */
-        var $el = $sub.parents("form"),
-            cfgs = $sub.data("pat-inject");
+        const $el = $sub.parents("form");
+        const cfgs = $sub.data("pat-inject");
 
         // store the params of the subform in the config, to be used by history
-        $(cfgs).each(function (i, v) {
+        $(cfgs).each((i, v) => {
             v.params = $.param($sub.serializeArray());
         });
 
@@ -266,14 +219,14 @@ const inject = {
         } catch (e) {
             log.error("patterns-inject-triggered", e);
         }
-        inject.execute(cfgs, $el);
+        this.execute(cfgs, $el);
     },
 
-    extractConfig: function inject_extractConfig($el, opts) {
+    extractConfig($el, opts) {
         opts = $.extend({}, opts);
 
-        var cfgs = parser.parse($el, opts, true);
-        cfgs.forEach(function inject_extractConfig_each(cfg) {
+        const cfgs = parser.parse($el, opts, true);
+        cfgs.forEach((cfg) => {
             // opts and cfg have priority, fallback to href/action
             cfg.url =
                 opts.url ||
@@ -284,17 +237,18 @@ const inject = {
                 "";
 
             // separate selector from url
-            var urlparts = cfg.url.split("#");
+            const urlparts = cfg.url.split("#");
             cfg.url = urlparts[0];
-
-            // if no selector, check for selector as part of original url
-            var defaultSelector = (urlparts[1] && "#" + urlparts[1]) || "body";
 
             if (urlparts.length > 2) {
                 log.warn("Ignoring additional source ids:", urlparts.slice(2));
             }
 
-            cfg.defaultSelector = cfg.defaultSelector || defaultSelector;
+            if (!cfg.defaultSelector) {
+                // if no selector, check for selector as part of original url
+                cfg.defaultSelector =
+                    (urlparts[1] && "#" + urlparts[1]) || "body";
+            }
             if (cfg.delay) {
                 try {
                     cfg.delay = utils.parseTime(cfg.delay);
@@ -308,30 +262,30 @@ const inject = {
         return cfgs;
     },
 
-    elementIsDirty: function (m) {
+    elementIsDirty(m) {
         /* Check whether the passed in form element contains a value.
          */
-        var data = $.map(m.find(":input:not(select)"), function (i) {
-            var val = $(i).val();
+        const data = $.map(m.find(":input:not(select)"), (i) => {
+            const val = $(i).val();
             return Boolean(val) && val !== $(i).attr("placeholder");
         });
         return $.inArray(true, data) !== -1;
     },
 
-    askForConfirmation: function inject_askForConfirmation(cfgs) {
+    askForConfirmation(cfgs) {
         /* If configured to do so, show a confirmation dialog to the user.
          * This is done before attempting to perform injection.
          */
-        var should_confirm = false,
-            message;
+        let should_confirm = false;
+        let message;
 
-        _.each(cfgs, function (cfg) {
-            var _confirm = false;
+        _.each(cfgs, (cfg) => {
+            let _confirm = false;
             if (cfg.confirm == "always") {
                 _confirm = true;
             } else if (cfg.confirm === "form-data") {
                 if (cfg.target != "none")
-                    _confirm = inject.elementIsDirty(cfg.$target);
+                    _confirm = this.elementIsDirty(cfg.$target);
             } else if (cfg.confirm === "class") {
                 if (cfg.target != "none")
                     _confirm = cfg.$target.hasClass("is-dirty");
@@ -349,7 +303,7 @@ const inject = {
         return true;
     },
 
-    ensureTarget: function inject_ensureTarget(cfg, $el) {
+    ensureTarget(cfg, $el) {
         /* Make sure that a target element exists and that it's assigned to
          * cfg.$target.
          */
@@ -364,13 +318,13 @@ const inject = {
                 log.error("Need target selector", cfg);
                 return false;
             }
-            cfg.$target = inject.createTarget(cfg.target);
+            cfg.$target = this.createTarget(cfg.target);
             cfg.$injected = cfg.$target;
         }
         return true;
     },
 
-    verifySingleConfig: function inject_verifySingleonfig($el, url, cfg) {
+    verifySingleConfig($el, url, cfg) {
         /* Verify one of potentially multiple configs (i.e. argument lists).
          *
          * Extract modifiers such as ::element or ::after.
@@ -386,17 +340,17 @@ const inject = {
         cfg.source = cfg.source || cfg.defaultSelector;
         cfg.target = cfg.target || cfg.defaultSelector;
 
-        if (!inject.extractModifiers(cfg)) {
+        if (!this.extractModifiers(cfg)) {
             return false;
         }
-        if (!inject.ensureTarget(cfg, $el)) {
+        if (!this.ensureTarget(cfg, $el)) {
             return false;
         }
-        inject.listenForFormReset(cfg);
+        this.listenForFormReset(cfg);
         return true;
     },
 
-    verifyConfig: function inject_verifyConfig(cfgs, $el) {
+    verifyConfig(cfgs, $el) {
         /* Verify and post-process all the configurations.
          * Each "config" is an arguments list separated by the &&
          * combination operator.
@@ -407,11 +361,11 @@ const inject = {
          * Verification for each cfg in the array needs to succeed.
          */
         return cfgs.every(
-            _.partial(inject.verifySingleConfig, $el, cfgs[0].url)
+            _.partial(this.verifySingleConfig.bind(this), $el, cfgs[0].url)
         );
     },
 
-    listenForFormReset: function (cfg) {
+    listenForFormReset(cfg) {
         /* if pat-inject is used to populate target in some form and when
          * Cancel button is pressed (this triggers reset event on the
          * form) you would expect to populate with initial placeholder
@@ -419,35 +373,33 @@ const inject = {
         if (cfg.target === "none")
             // Special case, we don't want to display any return value.
             return;
-        var $form = cfg.$target.parents("form");
+        const $form = cfg.$target.parents("form");
         if (
             $form.length !== 0 &&
             cfg.$target.data("initial-value") === undefined
         ) {
             cfg.$target.data("initial-value", cfg.$target.html());
-            $form.on("reset", function () {
+            $form.on("reset", () => {
                 cfg.$target.html(cfg.$target.data("initial-value"));
             });
         }
     },
 
-    extractModifiers: function inject_extractModifiers(cfg) {
+    extractModifiers(cfg) {
         /* The user can add modifiers to the source and target arguments.
          * Modifiers such as ::element, ::before and ::after.
          * We identifiy and extract these modifiers here.
          */
-        var source_re = /^(.*?)(::element)?$/,
-            target_re = /^(.*?)(::element)?(::after|::before)?$/,
-            source_match = source_re.exec(cfg.source),
-            target_match = target_re.exec(cfg.target),
-            targetMod,
-            targetPosition;
+        const source_re = /^(.*?)(::element)?$/;
+        const target_re = /^(.*?)(::element)?(::after|::before)?$/;
+        const source_match = source_re.exec(cfg.source);
+        const target_match = target_re.exec(cfg.target);
 
         cfg.source = source_match[1];
         cfg.sourceMod = source_match[2] ? "element" : "content";
         cfg.target = target_match[1];
-        targetMod = target_match[2] ? "element" : "content";
-        targetPosition = (target_match[3] || "::").slice(2); // position relative to target
+        const targetMod = target_match[2] ? "element" : "content";
+        const targetPosition = (target_match[3] || "::").slice(2); // position relative to target
 
         if (cfg.loadingClass) {
             cfg.loadingClass += " " + cfg.loadingClass + "-" + targetMod;
@@ -462,23 +414,22 @@ const inject = {
         return true;
     },
 
-    createTarget: function inject_createTarget(selector) {
+    createTarget(selector) {
         /* create a target that matches the selector
          *
          * XXX: so far we only support #target and create a div with
          * that id appended to the body.
          */
-        var $target;
         if (selector.slice(0, 1) !== "#") {
             log.error("only id supported for non-existing target");
             return null;
         }
-        $target = $("<div />").attr({ id: selector.slice(1) });
+        const $target = $("<div />").attr({ id: selector.slice(1) });
         $("body").append($target);
         return $target;
     },
 
-    stopBubblingFromRemovedElement: function ($el, cfgs, ev) {
+    stopBubblingFromRemovedElement($el, cfgs, ev) {
         /* IE8 fix. Stop event from propagating IF $el will be removed
          * from the DOM. With pat-inject, often $el is the target that
          * will itself be replaced with injected content.
@@ -488,51 +439,47 @@ const inject = {
          *
          * See: http://stackoverflow.com/questions/7114368/why-is-jquery-remove-throwing-attr-exception-in-ie8
          */
-        var s; // jquery selector
-        for (var i = 0; i < cfgs.length; i++) {
-            s = cfgs[i].target;
-            if ($el.parents(s).addBack(s) && !ev.isPropagationStopped()) {
+        for (const cfg of cfgs) {
+            const sel = cfg.target;
+            if ($el.parents(sel).addBack(sel) && !ev.isPropagationStopped()) {
                 ev.stopPropagation();
                 return;
             }
         }
     },
 
-    _performInjection: function ($el, $source, cfg, trigger, title) {
+    _performInjection(target, $el, $source, cfg, trigger, title) {
         /* Called after the XHR has succeeded and we have a new $source
          * element to inject.
          */
         if (cfg.sourceMod === "content") {
             $source = $source.contents();
         }
-        var $src;
+        let $src;
         // $source.clone() does not work with shived elements in IE8
         if (
             document.all &&
             document.querySelector &&
             !document.addEventListener
         ) {
-            $src = $source.map(function () {
-                return $(this.outerHTML)[0];
-            });
+            $src = $source.map((idx, el) => $(el.outerHTML)[0]);
         } else {
             $src = $source.safeClone();
         }
-        var $target = $(this),
-            $injected = cfg.$injected || $src;
 
-        $src.findInclusive("img").on("load", function () {
-            $(this).trigger("pat-inject-content-loaded");
+        $src.findInclusive("img").on("load", (e) => {
+            $(e.target).trigger("pat-inject-content-loaded");
         });
+
+        const $injected = cfg.$injected || $src;
         // Now the injection actually happens.
-        if (inject._inject(trigger, $src, $target, cfg)) {
-            inject._afterInjection($el, $injected, cfg);
+        if (this._inject(trigger, $src, $(target), cfg)) {
+            this._afterInjection($el, $injected, cfg);
         }
         // History support. if subform is submitted, append form params
-        var glue = "?";
+        const glue = cfg.url.indexOf("?") > -1 ? "&" : "?";
         if (cfg.history === "record" && "pushState" in history) {
             if (cfg.params) {
-                if (cfg.url.indexOf("?") > -1) glue = "&";
                 history.pushState(
                     { url: cfg.url + glue + cfg.params },
                     "",
@@ -543,20 +490,20 @@ const inject = {
             }
             // Also inject title element if we have one
             if (title)
-                inject._inject(trigger, title, $("title"), {
+                this._inject(trigger, title, $("title"), {
                     action: "element",
                 });
         }
     },
 
-    _afterInjection: function ($el, $injected, cfg) {
+    _afterInjection($el, $injected, cfg) {
         /* Set a class on the injected elements and fire the
          * patterns-injected event.
          */
         $injected
-            .filter(function () {
+            .filter((idx, el_) => {
                 // setting data on textnode fails in IE8
-                return this.nodeType !== TEXT_NODE;
+                return el_.nodeType !== TEXT_NODE;
             })
             .data("pat-injected", { origin: cfg.url });
 
@@ -570,18 +517,18 @@ const inject = {
                 .parent()
                 .trigger("patterns-injected", [cfg, $el[0], $injected[0]]);
         } else {
-            $injected.each(function () {
+            $injected.each((idx, el_) => {
                 // patterns-injected event will be triggered for each injected (non-text) element.
-                if (this.nodeType !== TEXT_NODE) {
-                    $(this)
+                if (el_.nodeType !== TEXT_NODE) {
+                    $(el_)
                         .addClass(cfg["class"])
-                        .trigger("patterns-injected", [cfg, $el[0], this]);
+                        .trigger("patterns-injected", [cfg, $el[0], el_]);
                 }
             });
         }
 
         if (cfg.scroll && cfg.scroll !== "none") {
-            var scroll_container = cfg.$target
+            let scroll_container = cfg.$target
                 .parents()
                 .addBack()
                 .filter(":scrollable");
@@ -590,25 +537,23 @@ const inject = {
                 : window;
 
             // default for scroll===top
-            var top = 0;
-            var left = 0;
+            let top = 0;
+            let left = 0;
 
             if (cfg.scroll !== "top") {
-                var scroll_target;
-                if (cfg.scroll === "target") {
-                    scroll_target = cfg.$target[0];
-                } else {
-                    scroll_target = $injected.filter(cfg.scroll)[0];
-                }
+                const scroll_target =
+                    cfg.scroll === "target"
+                        ? cfg.$target[0]
+                        : $injected.filter(cfg.scroll)[0];
 
                 // Get the reference element to which against we calculate
                 // the relative position of the target.
                 // In case of a scroll container of window, we do not have
                 // getBoundingClientRect method, so get the body instead.
-                var scroll_container_ref = scroll_container;
-                if (scroll_container_ref === window) {
-                    scroll_container_ref = document.body;
-                }
+                const scroll_container_ref =
+                    scroll_container === window
+                        ? document.body
+                        : scroll_container;
 
                 // Calculate absolute [¹] position difference between
                 // scroll_container and scroll_target.
@@ -656,8 +601,7 @@ const inject = {
     },
 
     async _onInjectSuccess($el, cfgs, ev) {
-        var sources$,
-            data = ev && ev.jqxhr && ev.jqxhr.responseText;
+        let data = ev && ev.jqxhr && ev.jqxhr.responseText;
         if (!data) {
             log.warn("No response content, aborting", ev);
             return;
@@ -666,11 +610,11 @@ const inject = {
             // Special case, we want to call something, but we don't want to inject anything
             data = "";
         }
-        $.each(cfgs[0].hooks || [], function (idx, hook) {
-            $el.trigger("pat-inject-hook-" + hook);
-        });
-        inject.stopBubblingFromRemovedElement($el, cfgs, ev);
-        sources$ = await inject.callTypeHandler(
+        $.each(cfgs[0].hooks || [], (idx, hook) =>
+            $el.trigger("pat-inject-hook-" + hook)
+        );
+        this.stopBubblingFromRemovedElement($el, cfgs, ev);
+        const sources$ = await this.callTypeHandler(
             cfgs[0].dataType,
             "sources",
             $el,
@@ -678,7 +622,7 @@ const inject = {
         );
         /* pick the title source for dedicated handling later
           Title - if present - is always appended at the end. */
-        var title;
+        let title;
         if (
             sources$ &&
             sources$[sources$.length - 1] &&
@@ -687,23 +631,22 @@ const inject = {
         ) {
             title = sources$[sources$.length - 1];
         }
-        cfgs.forEach(function (cfg, idx) {
-            function perform_inject() {
+        cfgs.forEach((cfg, idx1) => {
+            const perform_inject = () => {
                 if (cfg.target != "none")
-                    cfg.$target.each(function () {
-                        inject._performInjection.apply(this, [
+                    cfg.$target.each((idx2, target) => {
+                        this._performInjection(
+                            target,
                             $el,
-                            sources$[idx],
+                            sources$[idx1],
                             cfg,
                             ev.target,
-                            title,
-                        ]);
+                            title
+                        );
                     });
-            }
+            };
             if (cfg.processDelay) {
-                setTimeout(function () {
-                    perform_inject();
-                }, cfg.processDelay);
+                setTimeout(() => perform_inject(), cfg.processDelay);
             } else {
                 perform_inject();
             }
@@ -712,60 +655,82 @@ const inject = {
             // In case next-href is specified the anchor's href will
             // be set to it after the injection is triggered.
             $el.attr({ href: cfgs[0].nextHref.replace(/&amp;/g, "&") });
-            inject.destroy($el);
+            this.destroy($el);
         }
         $el.off("pat-ajax-success.pat-inject");
         $el.off("pat-ajax-error.pat-inject");
     },
 
-    _onInjectError: function ($el, cfgs, event) {
-        var explanation = "";
-        var timestamp = new Date();
-        if (event.jqxhr.status % 100 == 4) {
+    async _onInjectError($el, cfgs, event) {
+        let explanation = "";
+        let fallback;
+        const status = event.jqxhr.status;
+        const timestamp = new Date();
+        if (status % 100 == 4) {
             explanation =
                 "Sorry! We couldn't find the page to load. Please make a screenshot and send it to support. Thank you!";
-        } else if (event.jqxhr.status % 100 == 5) {
+        } else if (status % 100 == 5) {
             explanation =
                 "I am very sorry! There was an error at the server. Please make a screenshot and contact support. Thank you!";
-        } else if (event.jqxhr.status == 0) {
+        } else if (status == 0) {
             explanation =
                 "It seems, the server is down. Please make a screenshot and contact support. Thank you!";
         }
-        var msg_attr =
-            explanation +
-            " Status is " +
-            event.jqxhr.status +
-            " " +
-            event.jqxhr.statusText +
-            ", time was " +
-            timestamp +
-            ". You can click to close this.";
-        $("body").attr("data-error-message", msg_attr);
-        $("body").on("click", function () {
-            $("body").removeAttr("data-error-message");
-            window.location.href = window.location.href;
-        });
-        cfgs.forEach(function (cfg) {
+
+        const url_params = new URLSearchParams(window.location.search);
+
+        const fallback_url = document
+            .querySelector(`meta[name=pat-inject-${status}]`)
+            ?.getAttribute("content", false);
+        if (
+            fallback_url &&
+            url_params.get("pat-inject-errorhandler.off") === null
+        ) {
+            try {
+                const fallback_response = await fetch(fallback_url, {
+                    method: "GET",
+                });
+                fallback = document.createElement("html");
+                fallback.innerHTML = await fallback_response.text();
+                fallback = fallback.querySelector("body");
+            } catch {
+                // fallback to standard error message and ignore.
+            }
+        }
+
+        // clean up
+        cfgs.forEach((cfg) => {
             if ("$injected" in cfg) cfg.$injected.remove();
         });
         $el.off("pat-ajax-success.pat-inject");
         $el.off("pat-ajax-error.pat-inject");
+
+        if (fallback) {
+            document.body.innerHTML = fallback.innerHTML;
+        } else {
+            const msg_attr =
+                fallback ||
+                `${explanation} Status is ${status} ${event.jqxhr.statusText}, time was ${timestamp}. You can click to close this.`;
+            $("body").attr("data-error-message", msg_attr);
+            $("body").on("click", () => {
+                $("body").removeAttr("data-error-message");
+                window.location.href = window.location.href; // reload
+            });
+        }
     },
 
-    execute: function inject_execute(cfgs, $el) {
+    execute(cfgs, $el) {
         /* Actually execute the injection.
          *
          * Either by making an ajax request or by spoofing an ajax
          * request when the content is readily available in the current page.
          */
         // get a kinda deep copy, we scribble on it
-        cfgs = cfgs.map(function (cfg) {
-            return $.extend({}, cfg);
-        });
-        if (!inject.verifyConfig(cfgs, $el)) {
+        cfgs = cfgs.map((cfg) => $.extend({}, cfg));
+        if (!this.verifyConfig(cfgs, $el)) {
             return;
         }
-        if (!inject.askForConfirmation(cfgs)) {
+        if (!this.askForConfirmation(cfgs)) {
             return;
         }
         if ($el.data("pat-inject-triggered")) {
@@ -776,16 +741,14 @@ const inject = {
         // possibility for spinners on targets
         _.chain(cfgs)
             .filter(_.property("loadingClass"))
-            .each(function (cfg) {
+            .each((cfg) => {
                 if (cfg.target != "none")
                     cfg.$target.addClass(cfg.loadingClass);
             });
         // Put the execute class on the elem that has pat inject on it
         _.chain(cfgs)
             .filter(_.property("loadingClass"))
-            .each(function (cfg) {
-                $el.addClass(cfg.executingClass);
-            });
+            .each((cfg) => $el.addClass(cfg.executingClass));
 
         $el.on(
             "pat-ajax-success.pat-inject",
@@ -795,11 +758,8 @@ const inject = {
             "pat-ajax-error.pat-inject",
             this._onInjectError.bind(this, $el, cfgs)
         );
-        $el.on(
-            "pat-ajax-success.pat-inject pat-ajax-error.pat-inject",
-            function () {
-                $el.removeData("pat-inject-triggered");
-            }
+        $el.on("pat-ajax-success.pat-inject pat-ajax-error.pat-inject", () =>
+            $el.removeData("pat-inject-triggered")
         );
 
         if (cfgs[0].url.length) {
@@ -818,10 +778,10 @@ const inject = {
         }
     },
 
-    _inject: function inject_inject(trigger, $source, $target, cfg) {
+    _inject(trigger, $source, $target, cfg) {
         // action to jquery method mapping, except for "content"
         // and "element"
-        var method = {
+        const method = {
             contentbefore: "prepend",
             contentafter: "append",
             elementbefore: "before",
@@ -860,16 +820,16 @@ const inject = {
         return true;
     },
 
-    _sourcesFromHtml: function inject_sourcesFromHtml(html, url, sources) {
-        var $html = inject._parseRawHtml(html, url);
-        return sources.map(function inject_sourcesFromHtml_map(source) {
+    _sourcesFromHtml(html, url, sources) {
+        const $html = this._parseRawHtml(html, url);
+        return sources.map((source) => {
             if (source === "body") {
                 source = "#__original_body";
             }
             if (source === "none") {
                 return $("<!-- -->");
             }
-            var $source = $html.find(source);
+            const $source = $html.find(source);
 
             if ($source.length === 0) {
                 if (source != "title") {
@@ -877,8 +837,8 @@ const inject = {
                 }
             }
 
-            $source.find('a[href^="#"]').each(function () {
-                var href = this.getAttribute("href");
+            $source.find('a[href^="#"]').each((idx, el_) => {
+                const href = el_.getAttribute("href");
                 if (href.indexOf("#{1}") !== -1) {
                     // We ignore hrefs containing #{1} because they're not
                     // valid and only applicable in the context of
@@ -889,9 +849,9 @@ const inject = {
                 // this fragment.
                 if (href.length === 1) {
                     // Special case for top-of-page links
-                    this.href = url;
+                    el_.href = url;
                 } else if (!$source.find(href).length) {
-                    this.href = url + href;
+                    el_.href = url + href;
                 }
             });
             return $source;
@@ -907,12 +867,12 @@ const inject = {
         VIDEO: "data-pat-inject-rebase-src",
     },
 
-    _rebaseHTML: function inject_rebaseHTML(base, html) {
+    _rebaseHTML(base, html) {
         if (html === "") {
             // Special case, source is none
             return "";
         }
-        var $page = $(
+        const $page = $(
             html
                 .replace(
                     /(\s)(src\s*)=/gi,
@@ -924,11 +884,11 @@ const inject = {
             .parent();
 
         $page
-            .find(Object.keys(inject._rebaseAttrs).join(","))
-            .each(function () {
-                var $this = $(this),
-                    attrName = inject._rebaseAttrs[this.tagName],
-                    value = $this.attr(attrName);
+            .find(Object.keys(this._rebaseAttrs).join(","))
+            .each((idx, el_) => {
+                const $el_ = $(el_);
+                const attrName = this._rebaseAttrs[el_.tagName];
+                let value = $el_.attr(attrName);
 
                 if (
                     value &&
@@ -938,13 +898,13 @@ const inject = {
                     value.slice(0, 11) !== "javascript:"
                 ) {
                     value = utils.rebaseURL(base, value);
-                    $this.attr(attrName, value);
+                    $el_.attr(attrName, value);
                 }
             });
         // XXX: IE8 changes the order of attributes in html. The following
         // lines move data-pat-inject-rebase-src to src.
-        $page.find("[data-pat-inject-rebase-src]").each(function () {
-            var $el = $(this);
+        $page.find("[data-pat-inject-rebase-src]").each((id, el_) => {
+            const $el = $(el_);
             $el.attr("src", $el.attr("data-pat-inject-rebase-src")).removeAttr(
                 "data-pat-inject-rebase-src"
             );
@@ -956,12 +916,12 @@ const inject = {
             .trim();
     },
 
-    _parseRawHtml: function inject_parseRawHtml(html, url) {
+    _parseRawHtml(html, url) {
         url = url || "";
 
         // remove script tags and head and replace body by a div
-        var title = html.match(/\<title\>(.*)\<\/title\>/);
-        var clean_html = html
+        const title = html.match(/\<title\>(.*)\<\/title\>/);
+        let clean_html = html
             .replace(/<script\b[^<]*(?:(?!<\/script>)<[^<]*)*<\/script>/gi, "")
             .replace(/<head\b[^<]*(?:(?!<\/script>)<[^<]*)*<\/head>/gi, "")
             .replace(/<body([^>]*?)>/gi, '<div id="__original_body">')
@@ -970,11 +930,11 @@ const inject = {
             clean_html = title[0] + clean_html;
         }
         try {
-            clean_html = inject._rebaseHTML(url, clean_html);
+            clean_html = this._rebaseHTML(url, clean_html);
         } catch (e) {
             log.error("Error rebasing urls", e);
         }
-        var $html = $("<div/>").html(clean_html);
+        const $html = $("<div/>").html(clean_html);
         if ($html.children().length === 0) {
             log.warn(
                 "Parsing html resulted in empty jquery object:",
@@ -985,65 +945,61 @@ const inject = {
     },
 
     // XXX: hack
-    _initAutoloadVisible: function inject_initAutoloadVisible($el, cfgs) {
+    _initAutoloadVisible($el, cfgs) {
         if ($el.data("pat-inject-autoloaded")) {
             // ignore executed autoloads
             return false;
         }
-        var $scrollable = $el.parents(":scrollable"),
-            checkVisibility;
+        const $scrollable = $el.parents(":scrollable");
 
         // function to trigger the autoload and mark as triggered
-        function trigger(event) {
+        const trigger = (event) => {
             if ($el.data("pat-inject-autoloaded")) {
                 return false;
             }
             $el.data("pat-inject-autoloaded", true);
-            inject.onTrigger.apply($el[0], []);
+            this.onTrigger({ target: $el[0] });
             event && event.preventDefault();
             return true;
-        }
+        };
         $el.click(trigger);
 
         // Use case 1: a (heigh-constrained) scrollable parent
         if ($scrollable.length) {
             // if scrollable parent and visible -> trigger it
             // we only look at the closest scrollable parent, no nesting
-            checkVisibility = utils.debounce(
-                function inject_checkVisibility_scrollable() {
-                    if (
-                        $el.data("patterns.autoload") ||
-                        !$.contains(document, $el[0])
-                    ) {
-                        return false;
-                    }
-                    if (!$el.is(":visible")) {
-                        return false;
-                    }
-                    // check if the target element still exists. Otherwise halt and catch fire
-                    var target = (
-                        $el.data("pat-inject")[0].target ||
-                        cfgs[0].defaultSelector
-                    ).replace(/::element/, "");
-                    if (target && target !== "self" && $(target).length === 0) {
-                        return false;
-                    }
-                    var reltop =
-                            $el.safeOffset().top -
-                            $scrollable.safeOffset().top -
-                            1000,
-                        doTrigger = reltop <= $scrollable.innerHeight();
-                    if (doTrigger) {
-                        // checkVisibility was possibly installed as a scroll
-                        // handler and has now served its purpose -> remove
-                        $($scrollable[0]).off("scroll", checkVisibility);
-                        $(window).off("resize.pat-autoload", checkVisibility);
-                        return trigger();
-                    }
+            // Check visibility for scrollable
+            const checkVisibility = utils.debounce(() => {
+                if (
+                    $el.data("patterns.autoload") ||
+                    !$.contains(document, $el[0])
+                ) {
                     return false;
-                },
-                100
-            );
+                }
+                if (!$el.is(":visible")) {
+                    return false;
+                }
+                // check if the target element still exists. Otherwise halt and catch fire
+                const target = (
+                    $el.data("pat-inject")[0].target || cfgs[0].defaultSelector
+                ).replace(/::element/, "");
+                if (target && target !== "self" && $(target).length === 0) {
+                    return false;
+                }
+                const reltop =
+                        $el.safeOffset().top -
+                        $scrollable.safeOffset().top -
+                        1000,
+                    doTrigger = reltop <= $scrollable.innerHeight();
+                if (doTrigger) {
+                    // checkVisibility was possibly installed as a scroll
+                    // handler and has now served its purpose -> remove
+                    $($scrollable[0]).off("scroll", checkVisibility);
+                    $(window).off("resize.pat-autoload", checkVisibility);
+                    return trigger();
+                }
+                return false;
+            }, 100);
             if (checkVisibility()) {
                 return true;
             }
@@ -1052,45 +1008,40 @@ const inject = {
             $(window).on("resize.pat-autoload", checkVisibility);
         } else {
             // Use case 2: scrolling the entire page
-            checkVisibility = utils.debounce(
-                function inject_checkVisibility_not_scrollable() {
-                    if ($el.parents(":scrollable").length) {
-                        // Because of a resize the element has now a scrollable parent
-                        // and we should reset the correct event
-                        $(window).off(".pat-autoload", checkVisibility);
-                        return inject._initAutoloadVisible($el);
-                    }
-                    if ($el.data("patterns.autoload")) {
-                        return false;
-                    }
-                    if (!$el.is(":visible")) {
-                        return false;
-                    }
-                    if (!utils.elementInViewport($el[0])) {
-                        return false;
-                    }
-                    // check if the target element still exists. Otherwise halt and catch fire
-                    var target = (
-                        $el.data("pat-inject")[0].target ||
-                        cfgs[0].defaultSelector
-                    ).replace(/::element/, "");
-                    if (target && target !== "self" && $(target).length === 0) {
-                        return false;
-                    }
+            // Check visibility for non-scrollable
+            const checkVisibility = utils.debounce(() => {
+                if ($el.parents(":scrollable").length) {
+                    // Because of a resize the element has now a scrollable parent
+                    // and we should reset the correct event
                     $(window).off(".pat-autoload", checkVisibility);
-                    return trigger();
-                },
-                100
-            );
+                    return this._initAutoloadVisible($el);
+                }
+                if ($el.data("patterns.autoload")) {
+                    return false;
+                }
+                if (!$el.is(":visible")) {
+                    return false;
+                }
+                if (!utils.elementInViewport($el[0])) {
+                    return false;
+                }
+                // check if the target element still exists. Otherwise halt and catch fire
+                const target = (
+                    $el.data("pat-inject")[0].target || cfgs[0].defaultSelector
+                ).replace(/::element/, "");
+                if (target && target !== "self" && $(target).length === 0) {
+                    return false;
+                }
+                $(window).off(".pat-autoload", checkVisibility);
+                return trigger();
+            }, 100);
             if (checkVisibility()) {
                 return true;
             }
             // https://github.com/w3c/IntersectionObserver/tree/master/polyfill
             if (IntersectionObserver) {
-                var observer = new IntersectionObserver(checkVisibility);
-                $el.each(function (idx, el) {
-                    observer.observe(el);
-                });
+                const observer = new IntersectionObserver(checkVisibility);
+                $el.each((idx, el) => observer.observe(el));
             } else {
                 $(window).on(
                     "resize.pat-autoload scroll.pat-autoload",
@@ -1101,18 +1052,18 @@ const inject = {
         return false;
     },
 
-    _initIdleTrigger: function inject_initIdleTrigger($el, delay) {
+    _initIdleTrigger($el, delay) {
         // XXX TODO: handle item removed from DOM
-        var timeout = parseInt(delay, 10);
-        var timer;
+        const timeout = parseInt(delay, 10);
+        let timer;
 
-        function onTimeout() {
-            inject.onTrigger.apply($el[0], []);
+        const onTimeout = () => {
+            this.onTrigger({ target: $el[0] });
             unsub();
             clearTimeout(timer);
-        }
+        };
 
-        var onInteraction = utils.debounce(function onInteraction() {
+        const onInteraction = utils.debounce(() => {
             if (!document.body.contains($el[0])) {
                 unsub();
                 return;
@@ -1121,10 +1072,10 @@ const inject = {
             timer = setTimeout(onTimeout, timeout);
         }, timeout);
 
-        function unsub() {
-            ["scroll", "resize"].forEach(function (e) {
-                window.removeEventListener(e, onInteraction);
-            });
+        const unsub = () => {
+            ["scroll", "resize"].forEach((e) =>
+                window.removeEventListener(e, onInteraction)
+            );
             [
                 "click",
                 "keypress",
@@ -1132,16 +1083,14 @@ const inject = {
                 "mousemove",
                 "touchstart",
                 "touchend",
-            ].forEach(function (e) {
-                document.removeEventListener(e, onInteraction);
-            });
-        }
+            ].forEach((e) => document.removeEventListener(e, onInteraction));
+        };
 
         onInteraction();
 
-        ["scroll", "resize"].forEach(function (e) {
-            window.addEventListener(e, onInteraction);
-        });
+        ["scroll", "resize"].forEach((e) =>
+            window.addEventListener(e, onInteraction)
+        );
         [
             "click",
             "keypress",
@@ -1149,20 +1098,18 @@ const inject = {
             "mousemove",
             "touchstart",
             "touchend",
-        ].forEach(function (e) {
-            document.addEventListener(e, onInteraction);
-        });
+        ].forEach((e) => document.addEventListener(e, onInteraction));
     },
 
     // XXX: simple so far to see what the team thinks of the idea
-    registerTypeHandler: function inject_registerTypeHandler(type, handler) {
-        inject.handlers[type] = handler;
+    registerTypeHandler(type, handler) {
+        this.handlers[type] = handler;
     },
 
     async callTypeHandler(type, fn, context, params) {
         type = type || "html";
-        if (inject.handlers[type] && $.isFunction(inject.handlers[type][fn])) {
-            return await inject.handlers[type][fn].apply(context, params);
+        if (this.handlers[type] && $.isFunction(this.handlers[type][fn])) {
+            return await this.handlers[type][fn].bind(this)(...params);
         } else {
             return null;
         }
@@ -1170,23 +1117,16 @@ const inject = {
 
     handlers: {
         html: {
-            sources: function (cfgs, data) {
-                var sources = cfgs.map(function (cfg) {
-                    return cfg.source;
-                });
+            sources(cfgs, data) {
+                const sources = cfgs.map((cfg) => cfg.source);
                 sources.push("title");
-                return inject._sourcesFromHtml(data, cfgs[0].url, sources);
+                return this._sourcesFromHtml(data, cfgs[0].url, sources);
             },
         },
     },
 };
 
-$(document).on("patterns-injected.inject", function onInjected(
-    ev,
-    cfg,
-    trigger,
-    injected
-) {
+$(document).on("patterns-injected.inject", (ev, cfg, trigger, injected) => {
     /* Listen for the patterns-injected event.
      *
      * Remove the "loading-class" classes from all injection targets and
@@ -1208,7 +1148,7 @@ $(document).on("patterns-injected.inject", function onInjected(
     }
 });
 
-$(window).on("popstate", function (event) {
+$(window).on("popstate", (event) => {
     // popstate also triggers on traditional anchors
     if (!event.originalEvent.state && "replaceState" in history) {
         try {